# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/users": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Pessoa"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Pessoa",
                        "name": "pessoa",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Pessoa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Pessoa"
                        }
                    }
                }
            }
        },
        "/users/{cpf}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "CPF para fazer a busca da pessoa",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ]
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cpf para deletar uma Pessoa",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Endereco": {
            "type": "object",
            "properties": {
                "CEP": {
                    "type": "string"
                },
                "UF": {
                    "type": "string"
                },
                "bairro": {
                    "type": "string"
                },
                "cidade": {
                    "type": "string"
                },
                "complemento": {
                    "type": "string"
                },
                "enderecoPessoa": {
                    "type": "integer"
                },
                "logradouro": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                }
            }
        },
        "model.Pessoa": {
            "type": "object",
            "properties": {
                "CPF": {
                    "type": "string"
                },
                "RG": {
                    "type": "string"
                },
                "dataNascimento": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enderecos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Endereco"
                    }
                },
                "nome": {
                    "type": "string"
                },
                "sobrenome": {
                    "type": "string"
                },
                "telefones": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Telefone"
                    }
                }
            }
        },
        "model.Telefone": {
            "type": "object",
            "properties": {
                "ddd": {
                    "type": "integer"
                },
                "numero": {
                    "type": "string"
                },
                "telefonePessoa": {
                    "type": "integer"
                },
                "tipo": {
                    "type": "string"
                }
            }
        }
    }
}